+++ src/store/actions/taskDetailAction.js
@@ -29,25 +29,6 @@ export const getTaskDetails = (taskID, projectId) => {
   };
 };
 
-export const submitUploadApproveDesignIntent = async (formData, id, headers = {}) => {
-  const api = new Api();
-  const axiosInstance = await api.init({ headers });
-  let apiURL;
-  //https://pegadev.pg.com/prweb/api/ArtworkAgilityFile/v2/assignments/ASSIGN-WORKLIST PG-AAS-WORK U-5708!UPLOADAPPROVEDESIGNINTENT_FLOW/actions/UploadApproveDesignIntent
-  
-  apiURL = `${DEVURL}/v2/assignments/ASSIGN-WORKLIST ${id}!UPLOADAPPROVEDESIGNINTENT_FLOW/actions/UploadApproveDesignIntent`;
-
-  const submitUploadApproveDesignIntentData = await axiosInstance({
-    url: apiURL,
-    method: "PATCH",
-    data: formData,
-  });
-  // if (submitUploadApproveDesignIntentData?.data?.ID) {
-  //   App.dispatchToStore(editProjectAction(formData));
-  // }
-  return submitUploadApproveDesignIntentData;
-};
-
 export const saveDesignIntent = async (formData, headers = {}) => {
   const api = new Api();
   const axiosInstance = await api.init({ headers });
